Revert "Merge branch 'mitsuo' into aroncillo"
This reverts commit a622d6a80afde22462acae0996aae39e1647e9c8, reversing
changes made to a0af8f9a94f7e63102ef0ab279209ac24e555d03.
