+++ package.json
@@ -20,13 +20,10 @@
     "jss": "^10.8.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-google-login": "^5.2.2",
     "react-markdown": "^7.0.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
-    "react-social-login-buttons": "^3.6.0",
     "react-swipeable-views": "^0.14.0",
-    "react-use-googlelogin": "^0.12.5",
     "web-vitals": "^1.1.2",
     "xlsx": "^0.17.3"
   },
