+++ src/components/useForm.js
@@ -13,7 +13,6 @@ export function useForm(initialFieldValues, validateOnChange=false, validate) {
     const [values, setValues] = useState(initialFieldValues);
     const [errors, setErrors] = useState({});
 
-    /* The e param is a React's SyntheticEvent */
     const handleInputChange = e => {
         /* retrieve values from "Default" event */
         const { name, value } = e.target
