+++ src/pages/Administrador/GestionUsuarios/GestionUsuarios.js
@@ -198,7 +198,7 @@ export default function GestionUsuarios() {
                 </InputAdornment>
               )
             }}
-            sx={{ width: .75, visibility: "hidden" }}
+            sx={{ width: .75 }}
             onChange={handleSearch}
             type="search"
           />
