+++ src/pages/Administrador/MantenimientoDepartamento/GestionDepartamento.js
@@ -19,8 +19,6 @@ import departamentoService from '../../../services/departamentoService';
 import SearchIcon from '@mui/icons-material/Search';
 import EditOutlinedIcon from '@mui/icons-material/EditOutlined';
 import CloseIcon from '@mui/icons-material/Close';
-import * as employeeService from '../../../services/employeeService'
-import { Form } from '../../../components/useForm'
 
 const tableHeaders = [
     {
@@ -220,7 +218,7 @@ export default function GestionDepartamento() {
                         </InputAdornment>
                     )
                     }}
-                    sx={{ width: .75, visibility: "hidden" }}
+                    sx={{ width: .75 }}
                     onChange={handleSearch}
                     type="search"
                 />
