+++ src/pages/AsistenteSeccion/CargaDocente/AccordionDetailsHorarioProfesor.js
@@ -8,13 +8,10 @@ import { Grid, Divider, Avatar } from '@mui/material'
 import { Accordion, AccordionSummary, AccordionDetails } from '@mui/material'
 import { DT } from '../../../components/DreamTeam/DT'
 import { styled } from '@mui/material/styles';
-import { Controls } from '../../../components/controls/Controls'
 
-function GenerateRow(props) {
-    const { docente } = props
+function generateRow(docente) {
     return (
         <>
-            <Controls.Divider/>
             <Grid container>
                 <Grid item xs={1}>
                     <Avatar>
@@ -55,12 +52,6 @@ function GenerateRow(props) {
 
 export default function AccordionDetailsHorarioProfesor(props) {
     const { docentes } = props
-    const listDocenteClase = docentes.clase.map(docente => 
-        <GenerateRow key={docente.id} docente={docente} />
-    )
-    const listDocenteLab = docentes.lab.map(docente => 
-        <GenerateRow key={docente.id} docente={docente} />
-    )
     return (
         <Paper elevation={1} sx={{p: 1}}>
             <Typography
@@ -70,8 +61,7 @@ export default function AccordionDetailsHorarioProfesor(props) {
             >
                 Lista de Docentes de Clases
             </Typography>
-            {listDocenteClase}
-            <Controls.Divider/>
+            {docentes.clase.map(docente => generateRow(docente))}
             <Typography
                 variant="h4"
                 py="4px"
@@ -79,7 +69,7 @@ export default function AccordionDetailsHorarioProfesor(props) {
             >
                 Lista de Docentes de Practicas
             </Typography>
-            {listDocenteLab}
+            {docentes.lab.map(docente => generateRow(docente))}
         </Paper>
     )
 }
