+++ src/pages/Dev/TestIndex.js
@@ -4,22 +4,26 @@ import ReactMarkdown from 'react-markdown'
 
 const markdown = `# Index of testing pages in TeacherSoft
 
-Dev:
+- [Asistente De Seccion](${window.location.origin + '/AsistenteSeccion'})
+- [Administrador](${window.location.origin + '/GestionUsuariosForm'})
+- [employee](${window.location.origin + '/employee'})
+- [login screen](${window.location.origin + '/login'})
+- [DreamTeam components showcase](${window.location.origin + '/showcase'})
+- [here](${window.location.href})
+- [Gestion Usuarios](${window.location.origin + '/GestionDeUsuarios'})
+- [Gestión carga cursos](${window.location.origin + '/GestionCargaCursos'})
 
-- [Employees.js](${window.location.origin + '/admin/employees'})
-- [Showcase.js](${window.location.origin + '/admin/showcase'})
-- [Login.js](${window.location.origin + '/login'})
-- [TestIndex.js](${window.location.href})
+~~~C
+# include <stdio.h>
 
-Admin:
+int main() {
+    printf("hola mundo!\n");
+    return 0;
+}
+~~~
 
-- [GestionUsuarios.js](${window.location.origin + '/admin/mantenimiento/usr'})
-- [GestionDepartamento.js](${window.location.origin + '/admin/mantenimiento/dep'})
-- [GestionSeccion.js](${window.location.origin + '/admin/mantenimiento/sec'})
-
-Asistente de Seccion (as):
-
-> [react-markdown](https://github.com/remarkjs/react-markdown)
+- [just a linnk](https://reactjs.com)
+- [react-markdown](https://github.com/remarkjs/react-markdown)
 `
 
 /* Index page for pages being develop */
