+++ src/pages/General/UserPage.js
@@ -12,7 +12,7 @@ import GestionDepartamento from '../Administrador/MantenimientoDepartamento/Gest
 import GestionSeccion from '../Administrador/MantenimientoSeccion/GestionSeccion'
 import CargaDocente from '../AsistenteSeccion/CargaDocente/CargaDocente';
 // import GestionUsuariosForm from '../Administrador/GestionUsuariosForm';
-import ProtectedRoute from './ProtectedRoute';
+import ProtectedRoute from './RouterProtected';
 import Vacio from '../Dev/Vacio'
 import { Assistant } from '@mui/icons-material';
 import GestionUsuarios from '../Administrador/GestionUsuarios/GestionUsuarios';
@@ -31,48 +31,40 @@ const DrawerHeader = styled("div")(({ theme }) => ({
 export default function UserPage() {
   return (
     /* Content Body (aka. AppMain) (lo que tiene el fondito de la ardillita) */
-    <Box
-      component="main"
+    <Box 
+      component="main" 
       /* fill remainder of body */
       width={1}
       // flexGrow={1}   // unnecessary
       // bottom="0px"
-      p={2}
+      p={2} 
       overflow="auto"   // grow with content
       /* fondo y ardillita loca */
       // transform='translateZ(0)'
       sx={{
-        backgroundColor: "#ffffff",
+        backgroundColor:"#ffffff",
         // backgroundImage:'url("assets/img/ardillaloca.svg"), url("assets/img/rayaslocas.svg")',
-        backgroundImage: 'url("assets/img/fondoDT.svg")',
-        backgroundRepeat: 'no-repeat',
-        backgroundPosition: 'bottom right',
+        backgroundImage:'url("assets/img/fondoDT.svg")',
+        backgroundRepeat:'no-repeat',
+        backgroundPosition:'bottom right',
         // backgroundSize:'30%',
-        backgroundSize: 'contain',
+        backgroundSize:'contain',
       }}
     >
       {/* USER PAGE */}
       <DrawerHeader />  {/* SOLAMENTE UN DIV PARA HACER MARGIN TOP, 
                             codigo repetido en HeaderUser */}
       <Switch>
-
+        
         {/* admin */}
-        <ProtectedRoute exact path="/login" idRoles={[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]}>
-          <Login />
+        <ProtectedRoute exact path="/login" idRoles={[1,2,3,4,5,6,7,8,9,10]}>
+          <Login/>
         </ProtectedRoute>
 
         {/* DUCKTAPE */}
-        <ProtectedRoute exact path="/admin" idRoles={[1]}>
-          <Redirect to="/admin/mantenimiento/usr" />
-        </ProtectedRoute>
-
         <ProtectedRoute exact path="/admin/mantenimiento" idRoles={[1]}>
           <Redirect to="/admin/mantenimiento/usr" />
         </ProtectedRoute>
-
-        {/* Not yet listed */}
-        <ProtectedRoute exact path="/admin/mantenimiento/per" idRoles={[1]} component={Vacio} />
-        {/* ----end----- */}
 
         <ProtectedRoute exact path="/admin/mantenimiento/usr" idRoles={[1]} component={GestionUsuarios} />
         <ProtectedRoute exact path="/admin/mantenimiento/dep" idRoles={[1]} component={GestionDepartamento} />
@@ -82,10 +74,7 @@ export default function UserPage() {
         <ProtectedRoute exact path="/admin/employees" idRoles={[1]} component={Employees} />
 
         {/* as: asistente de seccion */}
-        <ProtectedRoute exact path="/as" idRoles={[1]}>
-          <Redirect to="/as/asignacionCarga/registroCursos" />
-        </ProtectedRoute>
-
+        <ProtectedRoute exact path="/as/asignacionCarga" idRoles={[2]} component={AsistenteSeccion} />
         <ProtectedRoute exact path="/as/asignacionCarga/registroCursos" idRoles={[2]} component={AsistenteSeccion} />
         <ProtectedRoute exact path="/as/asignacionCarga/cursos" idRoles={[2]} component={GestionCargaCursos} />
         <ProtectedRoute exact path="/as/asignacionCarga/registroCarga" idRoles={[2]} component={CargaDocente} />
