+++ src/pages/Login/LoginForm.js
@@ -4,14 +4,10 @@ import { Link, useHistory, useRouteMatch } from 'react-router-dom';
 import { Controls } from '../../components/controls/Controls'
 import { useForm, Form } from '../../components/useForm';
 import AuthService from '../../services/authService';
-import { GoogleLoginButton } from "react-social-login-buttons";
-import { GoogleLogin } from "react-google-login";
 import { useState } from 'react';
 import tokens from '../../services/tokens';
 import HeaderUser from '../../components/PageComponents/HeaderUser';
 import Vacio from '../Dev/Vacio';
-import axios from 'axios';
-import { useGoogleAuth } from './googleAuth';
 
 const initialFieldValues = {
   id: 0,
@@ -23,15 +19,7 @@ const initialFieldValues = {
 /*CODIGO -  */
 
 const Login = ({ setUser }) => {
-  const abrirGoogle= async () =>{
 
-    /* let res = await axios.get('http://localhost:8080/oauth2/authorization/google');
-
-    let data = res.data;
-    
-    }; 
-    /* windowObjectReference.close(); */
-  }
   //MANEJO DE ESTADOS
   //const [login, setLogin] = useState(null);
   const history = useHistory();
@@ -110,15 +98,6 @@ const Login = ({ setUser }) => {
     }
   }
 
-  const { signIn } = useGoogleAuth();
-  console.log(signIn)
-  /* const responseGoogle = (response) => {
-    console.log("aqui")
-    console.log(response);
-    signIn() ;
-    /* setUser(response.profileObj) ;
-  } */
-
   return (
     <Form onSubmit={handleSubmit} >
       <Controls.Input
@@ -155,26 +134,16 @@ const Login = ({ setUser }) => {
       <Typography align="center">
         o
       </Typography>
-      
-       {/* <GoogleLogin
-        clientId="626086626141-gclngcarehd8fhpacb2nrfq64mk6qf5o.apps.googleusercontent.com"
-        render={renderProps => (  */}
-          <Controls.Button
-                  variant="outlined"
-                  size='small'
-                  fullWidth
-                  text="Iniciar sesión con correo PUC"
-                  onClick={signIn}
-                />
-
-        {/* )}
-        buttonText="Iniciar sesión con correo PUCP"
-        onSuccess={responseGoogle}
-        onFailure={responseGoogle}
-        cookiePolicy={'single_host_origin'}
-
-        /> */}
-          
+      <a  href="localhost:8080/oauth2/authorization/google"
+         style={{ textDecoration: "none" }}
+      >
+        <Controls.Button
+          variant="outlined"
+          size='small'
+          fullWidth
+          text="Iniciar sesión con correo PUCP"
+        />
+      </a>
       <Typography paddingTop="20px" >
         <Link to="#" >
           Recuperar contraseña
