+++ src/pages/Showcase/DTPage.js
@@ -3,7 +3,6 @@
 import React from 'react'
 import { Typography, Box } from '@mui/material'
 import { DT } from '../../components/DreamTeam/DT'
-import { Controls } from '../../components/controls/Controls'
 
 export default function DTPage() {
     return (
@@ -12,7 +11,7 @@ export default function DTPage() {
                 Etiqueta.  Hay que hacerla mas pequeña
             </Typography>
             <DT.Etiqueta
-                type="pendiente"       // error | info | success | warning | pendiente
+                type="info"       // error | info | success | warning
                 text="Pendiente"
             />
 
@@ -23,11 +22,6 @@ export default function DTPage() {
                 <Box bgcolor="cyan" width="100px" height="100px">
                 </Box>
             </DT.BorderBox>
-
-            <Typography>
-                {"Simple Divider.  OJO que esta en `Controls` y no en `DT`"}
-            </Typography>
-            <Controls.Divider/>
         </div>
     )
 }
