+++ src/pages/Showcase/Showcase.js
@@ -10,7 +10,6 @@ import TablePage from "./TablePage";
 import UtilPage from "./UtilPage";
 import DTPage from "./DTPage";
 import TestPage from "./TestPage";
-import { useGoogleAuth } from '../Login/googleAuth';
 
 /* static MUI CSS (export it if it's going to be used a lot) */
 const paperCSS = {
@@ -29,10 +28,8 @@ function CenteredTitle(props) {
 }
 
 export default function Employees() {
-            
     return (
         <>
-        
             <Typography variant="h1"  component="div">
                 Showcasing DreamTeam Components
             </Typography>
